feat(client/header): 헤더 부분 사용자 정보 컴포넌트
사용자 정보 컴포넌트 안에 context를 사용해서 userEmail의 존재 유무에 따라서 다른 컴포넌트를 랜더링한다.
