--- client/src/components/users/accountContainer/index.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import UserInfo from "./UserInfo";
-import LoginButton from "./LoginButton";
-
-const AccountContainer = () => {
-  return (
-    <>
-      <UserInfo />
-      <LoginButton />
-    </>
-  );
-};
-
-export default AccountContainer;
